Need to test later!

+++ src/amf/ngap-build.h
@@ -54,7 +54,7 @@ ogs_pkbuf_t *ngap_build_amf_configuration_transfer(
 
 ogs_pkbuf_t *ngap_build_path_switch_ack(amf_ue_t *amf_ue);
 ogs_pkbuf_t *ngap_build_path_switch_failure(
-    uint32_t ran_ue_ngap_id, uint32_t amf_ue_ngap_id,
+    uint32_t ran_ue_ngap_id, uint64_t amf_ue_ngap_id,
     NGAP_Cause_PR group, long cause);
 
 ogs_pkbuf_t *ngap_build_handover_command(ran_ue_t *source_ue);
